requirements.txt
pandas
psycopg2-binary
pyspark
sqlalchemy